Split out DomainCoreApi from SystemDomainApi
Similar the runtime api `Core` that every substrate runtime must
implement, `DomainCoreApi` is a runtime api every subspace domain
runtime must implement.
